move more functions to model_functions
